UI: Полный редизайн мессенджера — тёмная тема, вкладки, пузырьки сообщений, автопрокрутка, правка входа/регистрации

+++ Models/TaskItems.cs
@@ -1,10 +1,8 @@
-// C:\Users\St\MessengerWithTasks\MessengerApp\Models\TaskItem.cs
-// Модель задачи
 namespace MessengerApp.Models
 {
     public class TaskItem
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string Title { get; set; } = string.Empty;
         public bool IsDone { get; set; }
     }
